requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0
boto3==1.34.144
razorpay==1.4.1
slowapi==0.1.9
apscheduler==3.10.4
httpx==0.27.0
email-validator==2.2.0
python-dotenv==1.0.1
aiofiles==23.2.1
